requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
apscheduler==3.10.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2